requirements.ml.txt
jupyter
numpy
pandas
scikit-surprise